refactor(core): extract query analysis + planning to rag_query (#478)
Closes #478. Fourth slice of the rag_core decomposition series (after
rag_retrieval PR-H1a/b, rag_verifier PR-J1, rag_answer PR-J2).

Moves 15 functions from rag_core.py:1318-1796 (~479 LOC) to new
rag_query.py:
- Query inspection: is_metadata_ambiguous, has_implicit_reference,
  has_comparison_request, extract_requested_agencies
- Conversation state helpers: active_state_terms, active_state_size,
  inject_entities_into_query, make_context_resolution
- Main entries: resolve_conversation_context, analyze_query
- Metadata helpers: comparison_targets_for_analysis,
  summarize_metadata_match, metadata_resolution_diagnostics
- Planner: query_type_default_top_k, make_plan

Cross-module impact handled
---------------------------
- comparison_targets_for_analysis is late-imported by rag_retrieval
  and rag_answer via `from rag_core import ...`. The re-export pattern
  in rag_core keeps that working — no changes needed in
  rag_retrieval/rag_answer.
- Orchestration (_phase_analyze, _phase_retrieve_loop) re-imports
  analyze_query and make_plan from rag_core (still works).

Dependency handling
-------------------
Module-level imports in rag_query (all leaves, no path back to
rag_core):
- korean_lexicon (IMPLICIT_REFERENCE_PATTERNS, STOPWORDS, TOPIC_KEYWORDS)
- rag_conversation_state (CONTEXT_RESOLUTION_THRESHOLD)
- rag_pipeline_presets (DEFAULT_RAG_PIPELINE_NAME, RRF_K,
  VALID_BM25_STOPWORD_PROFILES, VALID_RETRIEVAL_BACKENDS,
  VALID_RETRIEVAL_MODES, VALID_RRF_K_RANGE)
- text_normalize (normalize_text)

Late-imports (rag_core multi-surface utilities):
- is_metadata_ambiguous: metadata_ambiguity_details
- has_implicit_reference / has_comparison_request: normalize_entity
- extract_requested_agencies: ENTITY_RE, normalize_metadata_token,
  ordered_unique
- active_state_terms: coerce_string_list, ordered_unique
- resolve_conversation_context: coerce_string_list
- analyze_query: best_metadata_doc_scores, coerce_metadata_targets,
  match_metadata_targets, metadata_ambiguity_details,
  metadata_filters_from_matches, metadata_matches_for_stage,
  normalize_entity, ordered_unique, tokenize
- comparison_targets_for_analysis: ordered_unique
- metadata_resolution_diagnostics: coerce_string_list,
  compact_metadata_text, metadata_matches_for_stage, metadata_tokens,
  normalize_entity, ordered_unique
- query_type_default_top_k / make_plan: QUERY_TYPE_TOP_K_DEFAULTS

Re-exports from rag_core
------------------------
All 15 public names re-exported so any caller keeps working.

Cross-ref updates
-----------------
- CLAUDE.md: repository map entry for rag_query.py
- scripts/_governance.py: LOAD_BEARING_PATHS adds rag_query.py
- .github/pull_request_template.md: load-bearing hint includes rag_query.py
- ADR 0023 (HyDE) already pointed at rag_retrieval.retrieve_candidates;
  no cross-ref update needed.

LOC accounting
--------------
rag_core.py: 3,300 → 2,843 (-457)
rag_query.py: 0 → 597 (new, with 60-line module docstring + imports)

Cumulative across the series (PR-H1a + PR-H1b + PR-J1 + PR-J2 + PR-J3):
rag_core.py: 4,201 → 2,843 (-1,358 LOC, -32%)
New modules: rag_retrieval (~700), rag_verifier (~270),
rag_answer (~450), rag_query (~600). Each owns a cohesive
responsibility with leaf-only module-level imports + late imports
for rag_core utilities.

Verification
------------
Full local pytest 838 passed (1 skipped, 0 failed). Two rounds of
missing-import bugs caught and fixed during impl:
- metadata_filters_from_matches (used in analyze_query)
- QUERY_TYPE_TOP_K_DEFAULTS (used in make_plan)

The five regression gates stay green:
- tests/test_naive_baseline_ranking_invariance.py (ADR 0001 golden)
- tests/test_retrieval_loop_regression.py
- tests/test_langgraph_orchestrator_regression.py
- tests/test_conversation_state_*.py
- tests/test_governance.py (49 cases)
