+++ CLAUDE.md
@@ -48,6 +48,7 @@ Supporting:
 - `rag_retrieval.py` — retrieval pipeline extracted from `rag_core.py` across PR-H1a (issue #459) + PR-H1b (issue #461). Owns `retrieve_candidates` (candidate generation), the four similarity primitives (`embed_query_for_index`, `dense_similarity`, `lexical_similarity`, `metadata_similarity`), BM25 surface (`bm25_scores_for_index`, `get_or_build_bm25` + `_*` helpers), and the post-retrieval stack (`apply_fusion_and_reranking`, `apply_comparison_balance`, `reassemble_parent_sections`). Late-imports `tokenize` / `DEFAULT_EMBEDDING_MODEL` / `DEFAULT_HASH_DIM` / `embed_texts` / `hashing_embeddings` / `comparison_targets_for_analysis` / `normalize_regions` / `normalize_page_span` from `rag_core` to avoid circular imports — these helpers serve many non-retrieval rag_core call sites and stay there as canonical homes.
 - `rag_verifier.py` — verifier path extracted from `rag_core.py` (issue #465, PR-J1). Owns `verify_evidence` (main verifier + partial-topic grounding policy per ADR 0004), the topic-extraction helpers (`verification_topics`, `specific_topics`, `metadata_terms_for_verification`), the `EVIDENCE_BOUNDARY` constant + 3 instruction-pattern regexes, `neutralize_instruction_patterns` (ADR 0008 evidence-side defense), `evidence_text_for_verification`, `evidence_has_topic`, and the `PARTIAL_TOPIC_GROUNDING_*` policy constants. Direct imports: `korean_lexicon` (METADATA_EVIDENCE_LABELS / METADATA_GENERIC_TOKENS / TOPIC_KEYWORDS / VERIFICATION_INTENT_TOKENS) + `text_normalize` (expand_forms / normalize_text). Late-imports `normalize_metadata_token` / `metadata_tokens` / `ordered_unique` from `rag_core` so the module stays a leaf in the dependency graph. `rag_core` re-exports every public name + 3 constants so `tests/test_synthetic_judge.py` / `tests/test_prompt_injection_regression.py` / `scripts/llm_judge.py` / `eval/synthetic_judge.py` (which import `EVIDENCE_BOUNDARY` and `neutralize_instruction_patterns` from `rag_core`) keep working unchanged.
 - `rag_answer.py` — answer generation extracted from `rag_core.py` (issue #468, PR-J2). Owns the 20 functions that turn verified evidence into the ADR 0003 answer dict: `generate_answer` (main entry), `build_claims` / `build_comparison_claims` / `build_extract_claims`, `make_claim` / `make_citation` / `claim_target`, `answer_status` / `answer_status_reason` / `answer_query_type` / `answer_summary` / `answer_verification_reasons`, `build_insufficiency`, `render_answer_text`, `best_sentence` / `metadata_claim_sentences` / `metadata_field_requested` / `format_metadata_claim_value` / `sentence_has_verification_topic` / `select_supporting_evidence`. ADR 0003 dict contract (`schema_version: 2` literal) stays here; CLAUDE.md prohibition against parallel Pydantic models still applies. Direct imports: `korean_lexicon` (METADATA_CLAIM_*), `rag_answer_schema` (ANSWER_STATUS_* / ANSWER_SCHEMA_VERSION), `rag_verifier` (specific_topics / verification_topics / evidence_text_for_verification / PARTIAL_TOPIC_GROUNDING_REASON). Late-imports `sentence_split` / `tokenize` / `ordered_unique` / `normalize_regions` / `normalize_page_span` / `compact_metadata_text` from `rag_core` (multi-surface utilities). `rag_core` re-exports every public name so orchestration (`_phase_build_answer`) keeps working.
+- `rag_query.py` — query analysis + planning extracted from `rag_core.py` (issue #478, PR-J3). Owns the 15 functions that turn the raw user query into the `plan` dict retrieval consumes: `analyze_query` (main analyzer), `resolve_conversation_context` + `make_context_resolution` + 7 query-inspection / state helpers (`is_metadata_ambiguous` / `has_implicit_reference` / `has_comparison_request` / `extract_requested_agencies` / `active_state_terms` / `active_state_size` / `inject_entities_into_query`), `comparison_targets_for_analysis` (called via re-export by rag_retrieval / rag_answer), `summarize_metadata_match` / `metadata_resolution_diagnostics`, `query_type_default_top_k` / `make_plan`. Direct imports: `korean_lexicon` (IMPLICIT_REFERENCE_PATTERNS / STOPWORDS / TOPIC_KEYWORDS), `rag_conversation_state` (CONTEXT_RESOLUTION_THRESHOLD), `rag_pipeline_presets` (preset / validation constants), `text_normalize` (normalize_text). Late-imports many `rag_core` utilities (`tokenize` / `ordered_unique` / `normalize_entity` / `normalize_metadata_token` / `metadata_tokens` / `compact_metadata_text` / `match_metadata_targets` / `metadata_ambiguity_details` / `metadata_filters_from_matches` / `metadata_matches_for_stage` / `best_metadata_doc_scores` / `coerce_string_list` / `coerce_metadata_targets` / `ENTITY_RE` / `QUERY_TYPE_TOP_K_DEFAULTS`) so the module stays a leaf in the dependency graph. `rag_core` re-exports all 15 public names.
 - `rag_query_expansion.py` — `QueryExpander` Protocol + default `IdentityExpander` + opt-in `HyDEExpander` (issue #396, ADR 0023). Plugs in before the dense-embedding call in `rag_retrieval.retrieve_candidates`; BM25 / lexical / metadata paths consume `analysis.tokens` and remain invariant. Identity default preserves the ADR 0001 `naive_baseline` bit-identical golden.
 - `scripts/` — `build_index.py`, `update_readme_metrics.py`, `run_real_eval_delta.py`, etc.
 - `data/raw/` → `data/index/` → `outputs/` → `reports/` (pipeline artifacts).
